requirements.txt
jwt
cloudscraper
python-dotenv
anticaptchaofficial